create class for transformers
